Don't need the window resize callback
On most systems, the window size will match the framebuffer size. On
MacOS, the framebuffer size seems to be bigger than the window size
for retina displays. Because the renderer needs to fill the size of
the framebuffer, we only need to respond to framebuffer size
changes (ie. set glViewport to the new size). Window size is
irrelevant.