Add 'likely/unlikely' compiler hints + readable cplusplus values
C++20 adds [[likely]] and [[unlikely]] into the language. Previously
__builtin_expect could be used to provide that functionality. The
annoying thing is that the syntax of where it should be used
changed. I can make it portable for use in if statements, but not
switch statements. So that's all that supported right now (and all
that's needed right now! score!).

Additionally added proper defined to indicate the C++ version to be
used with RK_CPLUSPLUS. No more running about looking for the obscure values.