Properly support RTEK_MASTER_PROJECT
Unfortunately, since the introduction of RTEK_MASTER_PROJECT, none of
the compiler/linker config has been run! The check for the value of
the variable was wrong and was therefore never triggering. Thank
goodness it was found and now things are going to plan. As a
concequence, I realized that some of the values the configuration
needed weren't defined prior to use. Hence the file has been
reorganized.

This all started because /SUBSYSTEM:WINDOWS has been added in an
attempt to stop the initial console from swpaning on the Windows
platform. That work is still WIP, but the defines have been left in
this change since they are harmless are should stick around anyway.